fix: cold-start audit — all 21 findings (4 blocking, 6 major, 11 minor) (#152)
* fix(api): cold-start audit — forge 500s, scan-list hang, null lines, honest plan errors

- Widen ClassifiedTool.trust_score to float: fractional risk scores 500'd
  every /forge/search response (B-3, sibling of #147)
- /scans?scope=all: select slim columns instead of shipping NVARCHAR(MAX)
  JSON blobs for 1000 rows, compute findings_count via OPENJSON, parse
  MSSQL string-JSON metadata, add covering indexes (migration 009) (B-4)
- Finding.line is Optional[int]: CLI emits null for lineless findings and
  submission 422'd on most real scans, gating sigil explain (M-4)
- Auth0 /userinfo fallback: cache per-token, map 429/5xx to 503 instead of
  rewriting them as 401 "Invalid or expired token" (M-5)


* fix(cli): render all phases, emit valid JSON, drop self-inflicted provenance FPs

- phase_order now includes PromptInjection, SkillSecurity, InferenceSecurity —
  a prompt-injection-only scan printed CRITICAL RISK with zero findings shown (M-1)
- --format json emits one document ({summary, findings}); progress lines go
  to stderr; clone/pip/npm share the same output path; explain parses the
  new shape with legacy fallback (M-2)
- PROV-005 suppressed for sigil-made quarantine clones; PROV-006 only fires
  inside quarantine; explain hint shows real file-path usage; approve prints
  where the code lives (m-1, m-6)


* fix(dashboard): verify-email state, scan-history timeout, honest gating and copy

- AuthGuard third state: unverified sessions see a check-your-inbox screen
  instead of a silent /login loop; /api/auth/me distinguishes unverified (403)
  from unauthenticated (401); login page gains a create-account link (M-3)
- Scan History: 15s timeout + retryable error state instead of an infinite
  skeleton (B-4 front end)
- Free-plan policy settings disabled behind the gate; save errors surfaced (m-8)
- Community feed labeled as public/community with own-scans empty state (m-9)
- Annual badge computed from live plan prices (33%, was hardcoded 17%) (m-7)
- FP banner uses measured 70%→30% figures instead of unsourced 85% (m-10)
- Pro cards no longer promise unbuilt API access; dead OnboardingFlow /
  ApiKeySetupStep / 501 stub routes deleted, hardening test now guards that
  they stay deleted (M-6)


* fix(release,install,action): glibc floor, honest installer errors, JSON-driven CI summary

- Build linux binaries on ubuntu-22.04 / ubuntu-22.04-arm: glibc floor drops
  2.38/2.39 → 2.35, unbreaking Ubuntu 22.04 and Debian 12 (B-1)
- install.sh: stop discarding stderr (glibc and checksum failures surface
  distinctly), fall back to the releases/latest redirect on GitHub API rate
  limits, honor SIGIL_VERSION (B-1, m-4)
- action-entrypoint: build job summary from scan JSON instead of a
  never-written *_report.txt, derive phases from findings, fix 404ing
  nomark/sigil footer link to NOMARJ (m-3)
- Release notes: shasum -a 256 for macOS verification (m-5)


* docs: reconcile every published claim with what v1.3.6 actually does

- Remove Discovery Commands section and five other nonexistent commands;
  document token clearing as rm ~/.sigil/token (m-2)
- Real verdict thresholds and exit codes (no CLEAN, CRITICAL is
  evidence-gated); approve pins a ledger digest, nothing moves (m-1)
- Homebrew/npm out of Coming Soon; Team-tier scan/credit facts from the
  billing code; API-token instructions honest until key issuance ships (M-6)
- CI docs describe the action's real api-key input, not sigil-token (M-6)


* chore(release): 1.3.6 — cold-start audit fixes


---------

+++ api/routers/forge.py
@@ -77,7 +77,10 @@ class ClassifiedTool(BaseModel):
     ecosystem: str
     category: ToolCategory
     capabilities: list[ToolCapability]
-    trust_score: int
+    # float, not int: public_scans rows store fractional risk scores, so the
+    # derived trust score (100 - risk_score) is fractional too. Declaring int
+    # 500'd every /forge/search response (sibling of the #147 null-field bug).
+    trust_score: float
     verdict: str
     compatibility_signals: list[str]
     github_url: str | None = None
